requirements.txt
llama-cpp-python
orjson
numpy